Added Java 17 support to data and annotations bundles.
Data needed a dependency update for Nashorn since it was removed in Java
15. Annotations needed an updated version of Lombok and VM args updates
for Google's compile-test package to test the autogenerated code. Note
that I also removed the annotation testing lines that were checking the
defaults in every test. For some reason containsElementIn() now does an
exact test of the AST that wasn't a blocker before. However, testing for
defaults with every generation isn't needed since we do the exact
comparision with the expected source immediately after.

Signed-off-by: Jay Jay Billings <jayjaybillings@gmail.com>